requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via
    #   httpx
    #   openai
attrs==23.1.0
    # via
    #   automat
    #   service-identity
    #   twisted
automat==22.10.0
    # via twisted
beautifulsoup4==4.12.2
    # via -r requirements.in
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.3
    # via rich-click
commonmark==0.9.1
    # via rich
constantly==23.10.4
    # via twisted
cryptography==41.0.7
    # via
    #   pyopenssl
    #   scrapy
    #   service-identity
cssselect==1.2.0
    # via
    #   parsel
    #   scrapy
distro==1.9.0
    # via openai
filelock==3.13.1
    # via tldextract
h11==0.14.0
    # via httpcore
httpcore==1.0.2
    # via httpx
httpx==0.26.0
    # via openai
hyperlink==21.0.0
    # via twisted
idna==3.6
    # via
    #   anyio
    #   httpx
    #   hyperlink
    #   requests
    #   tldextract
incremental==22.10.0
    # via twisted
iniconfig==2.0.0
    # via pytest
itemadapter==0.8.0
    # via
    #   itemloaders
    #   scrapy
itemloaders==1.1.0
    # via scrapy
jmespath==1.0.1
    # via
    #   itemloaders
    #   parsel
lxml==4.9.3
    # via
    #   parsel
    #   scrapy
openai==1.6.1
    # via -r requirements.in
packaging==23.2
    # via
    #   parsel
    #   pytest
    #   scrapy
parsel==1.8.1
    # via
    #   itemloaders
    #   scrapy
pluggy==1.3.0
    # via pytest
protego==0.3.0
    # via scrapy
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.3.0
    # via service-identity
pycparser==2.21
    # via cffi
pydantic==2.5.3
    # via openai
pydantic-core==2.14.6
    # via pydantic
pydispatcher==2.0.7
    # via scrapy
pygments==2.13.0
    # via rich
pyopenssl==23.3.0
    # via scrapy
pytest==7.4.3
    # via -r requirements.in
python-dotenv==0.21.0
    # via -r requirements.in
queuelib==1.6.2
    # via scrapy
requests==2.31.0
    # via
    #   requests-file
    #   tldextract
requests-file==1.5.1
    # via tldextract
rich==12.5.1
    # via rich-click
rich-click==1.5.2
    # via -r requirements.in
ruff==0.1.5
    # via -r requirements.in
scrapy==2.11.0
    # via -r requirements.in
service-identity==23.1.0
    # via scrapy
six==1.16.0
    # via
    #   automat
    #   requests-file
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.5
    # via beautifulsoup4
tldextract==5.1.1
    # via scrapy
tqdm==4.66.1
    # via openai
twisted==22.10.0
    # via scrapy
typing-extensions==4.9.0
    # via
    #   openai
    #   pydantic
    #   pydantic-core
    #   twisted
urllib3==2.1.0
    # via requests
w3lib==2.1.2
    # via
    #   itemloaders
    #   parsel
    #   scrapy
zope-interface==6.1
    # via
    #   scrapy
    #   twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
